backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4
aiosmtplib==3.0.1
lxml==5.2.1
backports.zoneinfo;python_version<'3.9'
tzdata
slowapi
# SECURITY FIXES: Added for input validation and sanitization
bleach==6.1.0
email-validator==2.1.1
